BREAKING CHANGE: refactor skill system
Skills are now instances of Skill and ActiveSkill attached to the player capability. The player capability maps the skill's ID to a SkillWrapper containing the instance and level. Skill events are now contained and run in the SkillEvents class.

feat: add command to set and get primary skill

feat: add get, set, and clear to skill command

feat: add dash skill